refactor(enumerators): convert MaterialType to enum class with free functions
- Convert MaterialType from custom class to enum class : uint8_t
- Implement material_type_to_string and string_to_material_type free functions
- Update all usages in codebase and JSON serialization
- Update Lua bindings to use namespace constants
- Remove obsolete deriveClass for MaterialType

+++ src/creation/material.cpp
@@ -73,7 +73,7 @@ json::jnode_t &operator<<(json::jnode_t &lhs, const Material &rhs)
 {
     lhs.set_type(json::JTYPE_OBJECT);
     lhs["vnum"] << rhs.vnum;
-    lhs["type"] << rhs.type.toUInt();
+    lhs["type"] << static_cast<uint8_t>(rhs.type);
     lhs["name"] << rhs.name;
     lhs["article"] << rhs.article;
     lhs["worth"] << rhs.worth;
